Sunset/Sunrise offset
If you are like me and have a partner who complains that the lights aren't coming on soon enough, then this change will get you back in the good books.  Set an offset to Sunset or Sunrise up to a value of 60 mins that will allow you to fine tune the actions that are triggered by the sunset or sunrise events or the dark property.